+++ web/components/LessonScreen.tsx
@@ -18,6 +18,7 @@ import { formatClock } from '@/lib/telemetry-presentation'
 import { cn } from '@/lib/utils'
 import { AssignmentColumn } from './AssignmentColumn'
 import { ExerciseList } from './ExerciseList'
+import { LessonPrepPanel } from './LessonPrepPanel'
 import { useFleet } from './FleetProvider'
 import { formatElapsed } from './LessonStrip'
 import { StatusGlyph } from './StatusBadge'
@@ -154,6 +155,8 @@ function PreFlight({
           {withheld.map((drone) => drone.name).join(', ')}.
         </p>
       )}
+
+      <LessonPrepPanel drones={drones} book={book} />
 
       <AssignmentColumn drones={drones} book={book} />